:wrench: fix iteration bug

+++ src/huggingface.py
@@ -11,10 +11,14 @@ async def upload_to_huggingface(output_path):
         logging.error("No output path provided for uploading.")
         return
 
-    loop = asyncio.get_running_loop()
+    if not os.path.isfile(output_path):
+        logging.error(f"File {output_path} does not exist on the local file system.")
+        return
+
     try:
         print(f"😳 Uploading {output_path} to Hugging Face.")
-        result = await loop.run_in_executor(None, lambda: api.upload_file(
+        loop = asyncio.get_running_loop()
+        await loop.run_in_executor(None, lambda: api.upload_file(
             token=HFConfig.ACCESS_TOKEN,
             repo_id=HFConfig.REPO_ID,
             path_in_repo=os.path.basename(output_path),
@@ -24,8 +28,7 @@ async def upload_to_huggingface(output_path):
         print("🤗 Upload completed successfully.")
         if os.path.exists(output_path):
             os.remove(output_path)
-            print(f"🗑️  Deleted local file {output_path}")
-        return result
+            print(f"🗑️ Deleted local file {output_path}")
     except Exception as e:
         logging.error(f"Failed to upload {output_path} to Hugging Face: {e}")
-        raise  # Propagate exception to stop the process
+        raise
